refactor: extract CursorTooltip component
